Reguralarized for Linear Reservoir
Let the regularization parameter range from 1e-8 ------1e10 , found the results that all the NMSE curves get down at 1e6 and then reach a NMSE of  around 1.3.  Maybe because of when lambda is extremly large, it's like approximating using a constant and the variance of Lorenz itself is almost 1.3...........

Also integrated different curves in one figure.